==> src/04_probabilities.ipynb <==
import pickle
import altair as alt

import huspacy
import pandas as pd

from bertopic import BERTopic
from hdbscan import HDBSCAN
from bertopic.representation import KeyBERTInspired, MaximalMarginalRelevance
from sentence_transformers import SentenceTransformer

meta = pickle.load(open("../resources/meta.pkl", "rb"))
lemmatized = pickle.load(open("../resources/lemmatized.pkl", "rb"))
#pos = pickle.load(open("../resources/pos.pkl", "rb"))
#tokens =  pickle.load(open("../resources/tokenized.pkl", "rb"))
#doc_stop = pickle.load(open("../resources/no_stopword.pkl", "rb")) ### kevesebb stopszóval
doc_stop_2 = pickle.load(open("../resources/stopword_filtered.pkl", "rb")) ### több stop szóval, a no_stopword.pkl kiegészítve
all_docs = pickle.load(open("../resources/docs.pkl", "rb"))

prob_matrix = pd.read_csv("../results/prob_matrix.csv", sep=",")
prob_matrix["idx"] = list(range(prob_matrix.shape[0]))
prob_matrix.drop(columns=["25"], inplace=True)

import numpy as np

discourse_matrix = []

filenames = list(meta.values())
filenames = set([e.split(".txt")[0] for e in filenames])
for filename in filenames:
    file_meta = [(k,v) for k,v in meta.items() if filename in v]
    vals = [e[0] for e in file_meta]
    disc_prob_matrix = prob_matrix[prob_matrix["idx"].isin(vals)]
    disc_prob_matrix.drop(columns=["idx"], inplace=True)
    #save each discourse matrix to a separate csv file with the filename
    disc_prob_matrix.to_csv(f"../results/discourse_matrices/{filename}.csv", sep=",", encoding="UTF-8")
    m = np.mean(disc_prob_matrix.values, axis=0)
    discourse_matrix.append(m)

discourse_matrix = np.asarray(discourse_matrix)
discourse_df = pd.DataFrame(data=discourse_matrix,
                            index=list(range(discourse_matrix.shape[0])),
                            columns=list(range(25)))

discourse_df

topics=pd.read_csv("../results/bert_probabilities_merged_85_new_names.csv",sep=",")

topic_names = topics["CustomName"]
topic_names

discourse_df["diskurzus"] = list(range(discourse_matrix.shape[0]))
df_melted = pd.melt(
    discourse_df,
    id_vars=["diskurzus"],
    value_vars=list(range(25))
)

df_melted["topik"] = df_melted["variable"]
df_melted.drop(columns=["variable"], inplace=True)

df_melted.head()

df_melted["topik_nev"] = df_melted["topik"].apply(lambda x: topic_names[x])
df_melted.head()

df_melted.to_csv("../results/melted_discourse_probs.csv",sep=",", encoding="UTF-8")
