==> src/05_punchcard_discourse.ipynb <==
import altair as alt
import pandas as pd

df = pd.read_csv("../results/melted_discourse_probs.csv",sep=",")
df.drop(columns=["Unnamed: 0"], inplace=True)
df["value"] = df["value"].astype(float)
df["percent"] = df["value"] * 100
df.head(10)

alt.themes.enable("fivethirtyeight")

chart = (
    alt.Chart(df)
    .mark_circle()
    .encode(
        x=alt.X('diskurzus:N', axis=alt.Axis(labelAngle=-45)),
        y="topik:N",
        size=alt.Size("percent:Q",scale=alt.Scale(range=[0,1000]),title=None, legend=None),
        color=alt.Color("topik_nev:N",scale=alt.Scale(range=["blue", "red", "green", "yellow", "purple", "orange", "brown", "black", "grey", "violet", "cyan", "magenta", "olive", "maroon", "lime", "teal", "indigo", "darkgrey", "pink", "gold", "silver", "hotpink", "midnightblue", "darkorchid", "rosybrown"]),legend=alt.Legend(title="Topik")),
        tooltip=["topik", "percent"],
    )
).interactive()
chart.save("../vizs/punchcard_discourse.html")
