==> src/07_statistics.ipynb <==
import altair as alt
import pandas as pd

topics = pd.read_csv('../results/bert_probabilities_merged_85_new_names.csv', sep=',')
custom_names = topics['CustomName'].tolist()

discourse_probs = pd.read_csv('../results/discourse_probs.csv')

discourse_probs.drop(columns=['Unnamed: 0'], inplace=True)
discourse_probs  = discourse_probs * 100

discourse_probs.columns = custom_names

discourse_probs['discourse'] = range(1, 125)
discourse_probs


print(discourse_probs.columns)

for column in custom_names:
    chart = alt.Chart(discourse_probs, title=f"Topik: {column}").mark_bar(size=1).encode(
        x=alt.X('discourse:N', axis=alt.Axis(labelAngle=-45)),
        y=alt.Y(column, title=None),
        color=alt.value('yellow'),
        tooltip=[column],
    ).properties(
        width='container',
        height=500
    ).interactive()
    
    # Save the chart as an HTML file
    chart.save(f"../vizs/topic/{column}.html")
