==> src/08_count_topics.ipynb <==
import altair as alt
import pandas as pd
import pickle
import os
import re

ddir = "../data/combined_transcripts_freeConv"
meta = pickle.load(open("../resources/meta.pkl", "rb"))
topics = pd.read_csv('../results/bert_probabilities_merged_85_new_names.csv', sep=',')
topic_names = topics.set_index('Topic')['CustomName']
prob_matrix = pd.read_csv("../results/prob_matrix.csv", sep=",")
prob_matrix["idx"] = list(range(prob_matrix.shape[0]))
prob_matrix.drop(columns=["25"], inplace=True)
all_docs = pickle.load(open("../resources/docs.pkl", "rb"))
personality_df = pd.read_csv("../materials/commgame_quest_scores.csv", sep=',')

## Minden kísérleti alanynak kiszámoljuk, hogy átlagban a 25 témáról mennyit beszélt. A diskurzusban az ő beszédfordulói milyen arányben szólnak a 25 topik mindegxyikéről. 

import numpy as np

discourse_matrix_mordor = []
discourse_matrix_gondor = []
discourse_idx = []

filenames = list(meta.values())
filenames = set([e.split(".txt")[0] for e in filenames])
for filename in filenames:
    file_meta = [(k,v) for k,v in meta.items() if filename in v]
    vals = [e[0] for e in file_meta]
    vals_mordor = [e[0] for e in file_meta if "mordor" in e[1]]
    vals_gondor = [e[0] for e in file_meta if "gondor" in e[1]]
    match = re.search(r'pair(\d+)', filename)
    if match:
        numeric_value = int(match.group(1))
    pm = prob_matrix.drop(columns=["idx"])
    disc_prob_matrix_mordor = prob_matrix[prob_matrix["idx"].isin(vals_mordor)]
    disc_prob_matrix_gondor = prob_matrix[prob_matrix["idx"].isin(vals_gondor)]
    disc_prob_matrix_mordor.drop(columns=["idx"], inplace=True)
    disc_prob_matrix_gondor.drop(columns=["idx"], inplace=True)
    
    m_mordor = np.mean(disc_prob_matrix_mordor.values, axis=0)
    m_gondor = np.mean(disc_prob_matrix_gondor.values, axis=0)
    discourse_matrix_mordor.append(m_mordor)
    discourse_matrix_gondor.append(m_gondor)
    discourse_idx.append(numeric_value)

## A gondor és mondor alanyokat összerakjuk egy dataframebe. Az orszlopok a topikok, a sorok a beszélgetőpartnerek. Az értékek a valószínűségek. 2 plusz oszlop van: a páros száma és a labor neve (mordor vagy gondor).

discourse_matrix_mordor = np.asarray(discourse_matrix_mordor)
discourse_matrix_gondor = np.asarray(discourse_matrix_gondor)
discourse_df_mordor = pd.DataFrame(data=discourse_matrix_mordor,
                            index=list(range(discourse_matrix_mordor.shape[0])),
                            columns=list(range(25)))
discourse_df_mordor["PairNo"] = discourse_idx
discourse_df_gondor = pd.DataFrame(data=discourse_matrix_gondor,
                            index=list(range(discourse_matrix_gondor.shape[0])),
                            columns=list(range(25)))
discourse_df_gondor["PairNo"] = discourse_idx

discourse_df_mordor["LabName"] = "Mordor"
discourse_df_gondor["LabName"] = "Gondor"

discourse_df = pd.concat([discourse_df_mordor, discourse_df_gondor])
discourse_df

## Megnézzük, melyik 3 témáról beszéltek a legtöbbet az alanyok.

selected_columns = discourse_df.iloc[:,0:25]
top_3_columns = selected_columns.apply(lambda row: row.nlargest(3).index.tolist(), axis=1)
top_3_columns_df = pd.DataFrame(top_3_columns.tolist(), columns=['Top1', 'Top2', 'Top3'])
top_3_columns_df

## Megnézzük, hány darab topikról beszéltek több, mint 3%-ot az alanyok. Tetszés szerint változtathatjuk az arányt. Érdemes kísérletezni.

count_treshold = (discourse_df.iloc[:, 0:25] > 0.03).sum(axis=1)
count_treshold
